Get Api Data

+++ Assets/BitbankUnitySDK/Scripts/BitbankApi/Data/OrderData.cs
@@ -6,10 +6,10 @@ namespace BitbankUnitySDK
 {
     public class OrderData
     {
-        public int order_id;
+        public long order_id;
         public string pair;
-        public string side;
-        public string type;
+        public string side; // buy or sell
+        public string type; // limit or market
         public string start_amount;
         public string remaining_amount;
         public string executed_amount;
@@ -17,6 +17,6 @@ namespace BitbankUnitySDK
         public string average_price;
         public long ordered_at;
         public long? canceled_at;
-        public string status;
+        public string status; // UNFILLED, PARTIALLY_FILLED, FULLY_FILLED, CANCELED_UNFILLED, CANCELED_PARTIALLY_FILLED
     }
 }
